backend/requirements.txt
# --- AI & Orchestration ---
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-groq>=0.2.0
langchain-ollama>=0.2.0

# --- Web Interface ---
streamlit>=1.30.0

# --- API ---
fastapi>=0.109.0
uvicorn>=0.27.0

# --- Google Calendar API & Auth ---
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# --- Utilities ---
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0
